+++ src/errorHandler.ts
@@ -4,7 +4,7 @@ export function register() {
   if (process.env.SENTRY_PUBLIC_DSN) {
     Raven.config(<string> process.env.SENTRY_PUBLIC_DSN, {
       environment: process.env.NODE_ENV,
-      release: '1.9.0',
+      release: '1.10.0',
       serverName: process.env.BUILD_TARGET,
     }).install();
   }